libxfs: separate out userspace only files
To make checking that libxfs is identical between the kernel and
userspace code bases, lift all the userspace-only libxfs files into
a "support" subdirectory and build them there. Also, move all the
"kernel equivalent" XFS header files (e.g. include/xfs_mount.h) into
the support directory as well. This provides a clean separation of
libxfs from both kernel-specific and userspace-specific code bases,
and allows us to easily identify dependencies that must be fulfilled
for the libxfs code to work in userspace.

To make this work effectively, we have to enable sub-directory build
awareness in automake and slightly tweak the build rules to enable
libxfs/support to be built using the libxfs/Makefile.

Signed-off-by: Dave Chinner <dchinner@redhat.com>