spaceman: Free space mapping command
Add freespace mapping tool modelled on the xfs_db freesp command.
The advantage of this command over xfs_db is that it can be done
online and is coherent with concurrent modifications to the
filesystem.

This requires kernel extensions to the fiemap command to map free
space indexes.

The XFS_IOC_FIEMAPFS ioctl API definitions in this change are
derived from a patch supplied by Dhruvesh Rathore, Amey Ruikar and
Somdeep Dey.

Signed-off-by: Dave Chinner <dchinner@redhat.com>