libxfs-apply: allow commit range specification
Rather than having to manully run the script for every commit in a
series that needs to be applied, allow the commit ID to specify a
range of commits in the form or a git refspec.

Change the internal code to pull a commit at a time from the source
repository and applying it to the current repository. If guilt is in
use, this will result in a patch per commit, if guilt is not in use
it will aggregate all the changes into a single commit.

Also, fix the xfsprogs libxfs file filter match to exclude files
from the fs/xfs directory correctly.

Note: this pulls in a function from guilt to handle commit ids in a
sane manner.

Signed-off-by: Dave Chinner <david@fromorbit.com>